Rename confluent_kafka_input to input
